Revert "fix: feed Bresenham fractional position into controller to damp low-speed vel-fb oscillation"
This reverts commit 3c96fdcb24d6a7348f7616688e66b7f379eb669e.

+++ src/rp2040/pio.h
@@ -37,7 +37,7 @@ double compute_velocity_cmd(
     uint8_t  cmd_type,
     double   velocity_requested,
     double   abs_pos_requested,
-    double   abs_pos_achieved,
+    int32_t  abs_pos_achieved,
     uint8_t  enabled,
     uint32_t updated,
     uint32_t update_period_us,
